requirements.txt
configparser
python-telegram-bot
requests
python-dotenv
web3